Split feeder and serializer
